Added line numbers to the ouput of the parser to help debugging both in the compile and the render phase
Errors generated in the compile phase are now of the form:
{error, {Token, Reason, Line}}
Where Token is one of the atoms: include|apply|map|join|...

The compiled tokens are of the form:
{Token, Parameters, Line}

Parameters can be a single element or a tuple depending on the token.
E.g. include has a single element: the name of the template to be included
apply is a tuple like {FunToApply, Var}.